# tipu-api/openapi.yaml
openapi: 3.0.0
info:
  title: TIPU Academy API
  version: 1.0.0
  description: Online tutoring platform API with Firebase and Stripe integration
  contact:
    name: TIPU Academy Team
    email: support@tipuacademy.com
  license:
    name: MIT
servers:
  - url: http://localhost:8888
    description: Development server
  - url: https://api.tipuacademy.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase JWT authentication token
  schemas:
...
    Message:
      type: object
      required:
        - id
        - senderId
        - text
        - timestamp
      properties:
        id:
          type: string
          description: Message ID
          example: msg123
        senderId:
          type: string
          description: Sender user ID
          example: user123
        text:
          type: string
          description: Message text
          minLength: 1
          maxLength: 5000
          example: Hello, when is our next session?
        fileUrl:
          type: string
          format: uri
          description: Attached file URL (homework, documents)
          example: https://storage.googleapis.com/files/homework.pdf
        timestamp:
          type: string
          format: date-time
          description: Message timestamp
          example: "2025-11-23T10:30:00Z"
        read:
          type: boolean
          description: Read status
          example: false
    Conversation:
      type: object
      required:
        - id
        - participantIds
      properties:
        id:
          type: string
          description: Conversation ID
          example: conv123
        participantIds:
          type: array
          items:
            type: string
          description: Array of participant user IDs
          minItems: 2
          maxItems: 2
          example:
            - student123
            - tutor456
        lastMessage:
          type: object
          properties:
            text:
              type: string
              example: Thanks for the session!
            senderId:
              type: string
              example: student123
            timestamp:
              type: string
              format: date-time
              example: "2025-11-23T10:45:00Z"
        unreadCount:
          type: object
          description: Unread message count per user
          additionalProperties:
            type: number
          example:
            student123: 0
            tutor456: 2
        createdAt:
          type: string
          format: date-time
          example: "2025-11-20T09:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-11-23T10:45:00Z"
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Resource not found
        message:
          type: string
          description: Detailed error message
          example: The requested booking does not exist
        code:
          type: string
          description: Error code
          example: BOOKING_NOT_FOUND
paths:
...
  /v1/messages/conversations:
    get:
      tags:
        - Messages
      summary: Get user's conversations
      description: Retrieve all conversations for the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Maximum number of conversations to return
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of conversations to skip (for pagination)
      responses:
        "200":
          description: Conversations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Conversation"
              example:
                conversations:
                  - id: conv123
                    participantIds:
                      - student123
                      - tutor456
                    lastMessage:
                      text: Thanks for the session!
                      senderId: student123
                      timestamp: "2025-11-23T10:30:00Z"
                    unreadCount:
                      student123: 0
                      tutor456: 2
        "401":
          description: Unauthorized

    post:
      tags:
        - Messages
      summary: Create new conversation
      description: Create a new conversation between two users (e.g., student and tutor)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - participantIds
              properties:
                participantIds:
                  type: array
                  items:
                    type: string
                  minItems: 2
                  maxItems: 2
                  description: Array of user IDs to include in conversation
                  example:
                    - student123
                    - tutor456
      responses:
        "201":
          description: Conversation created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400":
          description: Invalid participant IDs or conversation already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
  /v1/messages/conversations/{id}:
    get:
      tags:
        - Messages
      summary: Get conversation messages
      description: Retrieve all messages in a specific conversation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Conversation ID
          example: conv123
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Maximum number of messages to return
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of messages to skip (for pagination)
      responses:
        "200":
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/Message"
              example:
                messages:
                  - id: msg123
                    senderId: student123
                    text: Hello, when is our next session?
                    timestamp: "2025-11-23T10:30:00Z"
                    read: true
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - User is not a participant in this conversation
        "404":
          description: Conversation not found
  /v1/messages/conversations/{id}/messages:
    post:
      tags:
        - Messages
      summary: Send message in conversation
      description: Send a new message in an existing conversation. Supports text messages and file attachments (homework, documents).
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Conversation ID
          example: conv123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: Message text content
                  minLength: 1
                  maxLength: 5000
                  example: Hello, when is our next session?
                fileUrl:
                  type: string
                  format: uri
                  description: Optional file attachment URL (homework, documents)
                  example: https://storage.googleapis.com/files/homework.pdf
      responses:
        "201":
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Invalid message content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - User is not a participant or student under 18 (chat restrictions)
        "404":
          description: Conversation not found
  /v1/messages/conversations/{id}/read:
    post:
      tags:
        - Messages
      summary: Mark messages as read
      description: Mark all messages in a conversation as read for the current user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Conversation ID
          example: conv123
      responses:
        "200":
          description: Messages marked as read successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Messages marked as read
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - User is not a participant in this conversation
        "404":
          description: Conversation not found